+++ python/tvm/backend/cuda/__init__.py
@@ -24,6 +24,7 @@ from tvm_ffi.libinfo import load_lib_ctypes
 from tvm.base import _LOADED_LIBS
 
 _LAZY_SUBMODULES = {
+    "codegen",
     "iket",
     "intrinsics",
     "lang",
@@ -70,6 +71,9 @@ def register_backend():
     for name, namespace in script_namespaces().items():
         builder_ir.register_script_namespace(name, namespace)
 
+    # script_namespaces() above pulls in ptx, which only imports the shared
+    # codegen layer -- not the device-helper modules. This import is the sole
+    # trigger that registers their codegens for codegen_cuda.cc to find.
     import_module(f"{__name__}.intrinsics")
     import_module(f"{__name__}.tile_primitive")
     import_module(f"{__name__}.target_tags")
@@ -106,6 +110,7 @@ def __getattr__(name: str):
 
 
 __all__ = [
+    "codegen",
     "iket",
     "intrinsics",
     "lang",
